+++ src/components/SingleDatePicker.jsx
@@ -32,6 +32,10 @@ const defaultProps = {
   focused: false,
   disabled: false,
   required: false,
+  showClearDate: false,
+  reopenPickerOnClearDate: false,
+  navPrev: null,
+  navNext: null,
 
   onDateChange() {},
   onFocusChange() {},
@@ -56,6 +60,7 @@ const defaultProps = {
   monthFormat: 'MMMM YYYY',
   phrases: {
     closeDatePicker: 'Close',
+    clearDate: 'Clear Date',
   },
 };
 
@@ -73,6 +78,7 @@ export default class SingleDatePicker extends React.Component {
     this.onChange = this.onChange.bind(this);
     this.onFocus = this.onFocus.bind(this);
     this.onClearFocus = this.onClearFocus.bind(this);
+    this.clearDate = this.clearDate.bind(this);
   }
 
   onChange(dateString) {
@@ -151,6 +157,14 @@ export default class SingleDatePicker extends React.Component {
     return typeof displayFormat === 'string' ? displayFormat : displayFormat();
   }
 
+  clearDate() {
+    const { onDateChange, reopenPickerOnClearDate, onFocusChange } = this.props;
+    onDateChange(null);
+    if (reopenPickerOnClearDate) {
+      onFocusChange({ focused: true });
+    }
+  }
+
   isBlocked(day) {
     const { isDayBlocked, isOutsideRange } = this.props;
     return isDayBlocked(day) || isOutsideRange(day);
@@ -186,6 +200,8 @@ export default class SingleDatePicker extends React.Component {
       numberOfMonths,
       orientation,
       monthFormat,
+      navPrev,
+      navNext,
       onPrevMonthClick,
       onNextMonthClick,
       withPortal,
@@ -223,6 +239,8 @@ export default class SingleDatePicker extends React.Component {
           hidden={!focused}
           initialVisibleMonth={initialVisibleMonth}
           onOutsideClick={onOutsideClick}
+          navPrev={navPrev}
+          navNext={navNext}
         />
 
         {withFullScreenPortal &&
@@ -248,7 +266,9 @@ export default class SingleDatePicker extends React.Component {
       focused,
       disabled,
       required,
+      showClearDate,
       date,
+      phrases,
       anchorDirection,
       withPortal,
       withFullScreenPortal,
@@ -281,6 +301,9 @@ export default class SingleDatePicker extends React.Component {
             disabled={disabled}
             required={required}
             showCaret={!withPortal && !withFullScreenPortal}
+            phrases={phrases}
+            onClearDate={this.clearDate}
+            showClearDate={showClearDate}
             dateValue={dateString}
             onChange={this.onChange}
             onFocus={this.onFocus}